P5: God-component splits — type relocation, page.tsx, project-overview decomposition
P5-1: Relocate PiAgentMessage/PiSubagentMessage/TerminalSession/PiSessionSummary
types from store/slices/terminal-sessions.ts → types/index.ts (688→620 LOC)
P5-2: Extract UpdateBanner + ErrorToasts from page.tsx inline → layout/app-chrome.tsx
(410→341 LOC)
P5-4: Decompose project-overview/index.tsx (664→427 LOC) — extracted 10 free
function sub-components into primitives.tsx (69 LOC) + sections.tsx (214 LOC)

All checks green: type-check:all, lint --max-warnings 0, compile, 430 tests.
